add check to make sure SDCard exists

+++ modules/restore.sh
@@ -2,6 +2,17 @@
 
 function restore {
 
+  a=$(fdisk -l |grep /dev/mmcblk0: | grep -P '\d+ (?=bytes)' -o)
+  #echo "$a - /dev/mmcblk0"
+  
+  b=$(fdisk -l |grep /dev/sdb: | grep -P '\d+ (?=bytes)' -o)
+  #echo "$b - /dev/sdb"
+  
+  if [ -z "$a" ] || [ -z "$b" ]; then
+      echo "no SDCard detected"
+      return 1
+  fi
+
   echo "restoring...."
   echo u > /proc/sysrq-trigger
   dd if=/dev/mmcblk0 bs=1M of=/dev/sdb
